separated functions and changed tests
