+++ tests/test-data.ts
@@ -1,3 +1,4 @@
+// table and json test data
 export const testCaseEmpty = {
 	md: "",
 	json: []
@@ -15,7 +16,8 @@ export const testCaseNoTableBody = {
 
 export const testCaseSmallTable = {
 	md: `| Test1 | Test2 |\n| --- | --- |\n| Value 1 | Value 2 |`,
-	json: [{Test1: "Value 1", Test2: "Value 2"}]
+	json: [{Test1: "Value 1", Test2: "Value 2"}],
+	lines: ["| Test1 | Test2 |", "| --- | --- |", "| Value 1 | Value 2 |"]
 };
 
 export const testCaseTableLarge = {
@@ -39,6 +41,13 @@ export const testCaseTableLarge = {
 			Test3: "",
 			Test4: "Value 3.4"
 		}
+	],
+	lines: [
+		"| Test1 | Test2 | Test3 | Test4 |",
+		"| --- | --- | --- | --- |",
+		"| Value 1.1 | Value 1.2 | Value 1.3 | Value 1.4 |",
+		"| Value 2.1 | Value 2.2 | Value 2.3 | |",
+		"| Value 3.1 | Value 3.2 | | Value 3.4 |"
 	]
 };
 
@@ -54,33 +63,6 @@ export const testCaseTableNDimension = {
 		}
 	]
 };
-
-export const noBodyTable = `| Test1 | Test |\n| --- | --- |`;
-export const testTable = `| Test1 | Test2 |\n| --- | --- |\n| Value 1 | Value 2 |`;
-
-export const testJson = [{Test1: "Value 1", Test2: "Value 2"}];
-export const testTableLarge = `| Test1 | Test2 | Test3 | Test4 |\n| --- | --- | --- | --- |\n| Value 1.1 | Value 1.2 | Value 1.3 | Value 1.4 |\n| Value 2.1 | Value 2.2 | Value 2.3 | |\n| Value 3.1 | Value 3.2 | | Value 3.4 |`;
-
-export const testJsonLarge = [
-	{
-		Test1: "Value 1.1",
-		Test2: "Value 1.2",
-		Test3: "Value 1.3",
-		Test4: "Value 1.4"
-	},
-	{
-		Test1: "Value 2.1",
-		Test2: "Value 2.2",
-		Test3: "Value 2.3",
-		Test4: ""
-	},
-	{
-		Test1: "Value 3.1",
-		Test2: "Value 3.2",
-		Test3: "",
-		Test4: "Value 3.4"
-	}
-];
 
 export const testTableNDimension = `| name | contact.mail | contact.phone.private | contact.phone.work | food[0].type | food[0].menus[0] | food[0].menus[1] | food[1].type | food[1].menus[0] | food[1].menus[1] |
 | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
@@ -134,27 +116,4 @@ export const testStringArray = [
 	"| Test1 | Test2 |",
 	"| --- | --- |",
 	"| Value 1 | Value 2 |"
-];
-
-export const testStringArrayLarge = [
-	"| Test1 | Test2 | Test3 | Test4 |",
-	"| --- | --- | --- | --- |",
-	"| Value 1.1 | Value 1.2 | Value 1.3 | Value 1.4 |",
-	"| Value 2.1 | Value 2.2 | Value 2.3 | |",
-	"| Value 3.1 | Value 3.2 | | Value 3.4 |"
-];
-
-export const keys = ["key1", "key2", "key3"];
-
-export const sameKeys = [
-	{key1: 1.1, key2: 1.2, key3: 1.3},
-	{key1: 2.1, key2: 2.2, key3: 2.3}
-];
-
-export const mixedKeysSingle = [{key1: 1.1}, {key2: 2.2}, {key3: 3.3}];
-
-export const mixedKeys = [
-	{key1: 1.1, key2: 1.2},
-	{key2: 2.2, key3: 2.3},
-	{key1: 3.1, key2: 3.2, key3: 3.3}
 ];
